+++ src/services/Formulas.php
@@ -100,10 +100,14 @@ class Formulas extends Component
             throw new SyntaxError('Tags are not allowed in a condition formula.');
         }
 
-        $cacheKey = 'formula:' . md5($formula) . '|params:' . md5(Json::encode($params));
+        $cacheKey = [
+            'formula' => md5($formula),
+            'params' => md5(Json::encode($params)),
+        ];
 
-        if (Craft::$app->getCache()->exists($cacheKey)) {
-            return (bool)Craft::$app->getCache()->get($cacheKey);
+        $cachedResult = Craft::$app->getCache()->get($cacheKey);
+        if ($cachedResult !== false) {
+            return $cachedResult === 'TRUE';
         }
 
         $twigCode = '{% if ';
@@ -112,10 +116,10 @@ class Formulas extends Component
 
         $template = $this->_twigEnv->createTemplate($twigCode, $name);
         $output = $template->render($params);
-        $result = ($output == 'TRUE');
-        Craft::$app->getCache()->set($cacheKey, $result);
 
-        return $result;
+        Craft::$app->getCache()->set($cacheKey, $output);
+
+        return $output === 'TRUE';
     }
 
     /**